[BE] refactor: RandomNumberGenerator 관련 패키지로 이동 및 참조 경로 수정 #774
