+++ backend/app/src/main/java/com/onair/hearit/app/explore/application/scorefactor/randomgenerator/DefaultRandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package com.onair.hearit.app.explore.application.scorefactor;
+package com.onair.hearit.app.explore.application.scorefactor.randomgenerator;
 
 import java.util.Random;
 import org.springframework.stereotype.Component;
