+++ backend/app/src/test/java/com/onair/hearit/app/explore/application/scorefactor/RandomScoreFactorTest.java
@@ -3,6 +3,8 @@ package com.onair.hearit.app.explore.application.scorefactor;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import com.onair.hearit.app.explore.application.scorefactor.randomgenerator.DefaultRandomNumberGenerator;
+import com.onair.hearit.app.explore.application.scorefactor.randomgenerator.RandomNumberGenerator;
 import com.onair.hearit.core.domain.Category;
 import com.onair.hearit.core.domain.Hearit;
 import com.onair.hearit.core.domain.Source;
